+++ src/biomass_relationship/nls/__init__.py
@@ -3,22 +3,13 @@ NLS module
 Used to explore how nitrogen deposition affects biomass through PCA-reduced variables
 """
 
-from .nls_model import (
-    NLSModel,
-    LinearModel,
-    AdditiveModel,
-    MultiplicativeModel,
-    MichaelisMentenModel,
-    ExponentialModel,
-    fit_nls_models,
-    compare_models,
-    ModelFitResult
-)
-
-from .analyzer import (
-    NLSAnalyzer,
-    quick_nls_analysis
-)
+from .nls_model import NLSModel
+from .linear_model import LinearModel
+from .additive_model import AdditiveModel
+from .multiplicative_model import MultiplicativeModel
+from .menten_model import MichaelisMentenModel
+from .exponential_model import ExponentialModel
+from .processor import fit_nls_models, compare_models
 
 __all__ = [
     'NLSModel',
@@ -28,9 +19,5 @@ __all__ = [
     'MichaelisMentenModel',
     'ExponentialModel',
     'fit_nls_models',
-    'compare_models',
-    'ModelFitResult',
-    'NLSAnalyzer',
-    'quick_nls_analysis'
+    'compare_models'
 ]
-
